Separate login functionality
Create LoginService.
